fix: 将 LKEmojiGroup 的原始类型从 UInt 更改为 UInt8

+++ Sources/LKEmoji/LKEmojiGroup.swift
@@ -1,4 +1,4 @@
-enum LKEmojiGroup: UInt {
+enum LKEmojiGroup: UInt8 {
     case neko = 15
     case luck = 10
     case onion = 5
